Updated coords.hpp, added calculateDistance method

+++ src/datasheet.hpp
@@ -2,12 +2,13 @@
 
 #include <string_view>
 #include <string>
+#include <cstdint>
 
 class Datasheet {
 public:
-    Datasheet(const std::string_view unitName, const unsigned int move, const unsigned int toughness,
-              const unsigned int save, const unsigned int wounds, const unsigned int leadership,
-              const unsigned int objectiveControl, const unsigned int modelsNum, const unsigned int cost);
+    Datasheet(const std::string_view unitName, const uint16_t move, const uint16_t toughness,
+              const uint16_t save, const uint16_t wounds, const uint16_t leadership,
+              const uint16_t objectiveControl, const uint16_t modelsNum, const uint16_t cost);
 
     uint16_t getModelsNum() const { return _modelsNum; }
     uint16_t getWounds() const { return _wounds; }
@@ -16,15 +17,15 @@ public:
 protected:
     std::string _unitName;
     // std::string image;
-    const unsigned int _move;
-    const unsigned int _toughness;
-    const unsigned int _save;
-    const unsigned int _wounds;
-    const unsigned int _leadership;
-    const unsigned int _objectiveControl;
+    const uint16_t _move;
+    const uint16_t _toughness;
+    const uint16_t _save;
+    const uint16_t _wounds;
+    const uint16_t _leadership;
+    const uint16_t _objectiveControl;
 
     // fields related to unit
-    const unsigned int _modelsNum;
-    const unsigned int _cost;
+    const uint16_t _modelsNum;
+    const uint16_t _cost;
     
 };
